+++ inspector/src/components/actionTabs/tabs/propertyGrids/animations/animationPropertyGridComponent.tsx
@@ -16,7 +16,7 @@ import { AnimationRange } from "babylonjs/Animations/animationRange";
 import { CheckBoxLineComponent } from "../../../lines/checkBoxLineComponent";
 import { Nullable } from "babylonjs/types";
 import { FloatLineComponent } from "../../../lines/floatLineComponent";
-import { TextLineComponent } from "../../../lines/textLineComponent";
+import { TextLineComponent } from "../../../../../sharedUiComponents/lines/textLineComponent";
 import { IAnimatable } from "babylonjs/Animations/animatable.interface";
 // import { AnimationCurveEditorComponent } from "../animations/animationCurveEditorComponent";
 // import { PopupComponent } from "../../../../popupComponent";
