+++ sharedUiComponents/lines/textLineComponent.tsx
